Crypto: Error API: ASAN Runtime error fix: Fix windows presubmit error

+++ lib/capi/c_error.cc
@@ -65,7 +65,7 @@ alcp_is_error(alc_error_t err)
     if (err_temp == 0)
         return false;
     else {
-        printf("Error code: %ld\n", err_temp);
+        printf("Error code: %ld\n", (long)err_temp);
         return true;
     }
 }
